Changed to cell Size and added water check!

+++ include/BoardOperations.hpp
@@ -14,7 +14,7 @@ namespace sas
     void multiplyPlants(sas::DynamicGrid &plantGrid, sas::StaticGrid &enviromentGrid, std::vector<std::unique_ptr<sas::Plant>> &plants) noexcept;
     void spawnWeed(StaticGrid &grid, const Matrix<Tile> &board, std::vector<std::unique_ptr<Plant>> &plants) noexcept;
     bool hasSpaceAgainstPlants(const Position &p, const std::vector<std::unique_ptr<Plant>> &plants, const DynamicGrid& plantGrid, const StaticGrid& envGrid) noexcept;
-    bool isNearWater(const Position &pos, const StaticGrid &waterCells, size_t maxRange) noexcept;
+    bool isNearWater(const Position &pos, const StaticGrid &waterCells, int maxRange) noexcept;
 
     void SetUpWater(std::vector<std::unique_ptr<Enviroment>> &waters, StaticGrid &grid) noexcept;
     void SetUpWaterNoise(std::vector<std::unique_ptr<sas::Enviroment>> &waters, sas::StaticGrid &grid, size_t seed) noexcept;
